refacto authslice + fix session cookie
