+++ front/app/components/CategoryList.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { AppDispatch, RootState } from "../store/store";
 import { incrementCount, decrementCount, setCount } from "../store/categoriesSlice";
 import Link from "next/link";
-import { deleteCategory, fetchCategories, saveProgress } from "../store/thunks";
+import { deleteCategory, fetchCategories, saveProgress } from "../store/categories.thunks";
 
 export default function CategoryList() {
   const dispatch = useDispatch<AppDispatch>();
