+++ front/app/dashboard/page.tsx
@@ -4,8 +4,8 @@ import { useAppDispatch, useAppSelector } from "../store/hooks";
 import CategoryForm from "../components/CategoryForm";
 import CategoryList from "../components/CategoryList";
 import { useRouter } from "next/navigation";
-import { logout } from "../store/authSlice";
-import { addCategory } from "../store/thunks";
+import { logout } from "../store/auth.thunks";
+import { addCategory } from "../store/categories.thunks";
 
 export default function Dashboard() {
   const dispatch = useAppDispatch();
